services/tasks/s3_exporter/requirements.txt
boto3==1.34.128
pandas==2.2.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
